services/agent-factory/pyproject.toml
[tool.poetry]
name = "agent-factory"
version = "0.1.0"
description = "Agent Factory service"
authors = ["OpenAI"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.8.0"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"